Initial work moving theme provider into a src folder
