+++ theme/doc/borderWidths.mdx
@@ -1,6 +1,6 @@
 import { mdx } from '@mdx-js/react';
 import { MdxLayout } from '~/layout';
-import { lightBaseTheme } from '../src/theme';
+import { lightBaseTheme } from '../src';
 import BorderWidthsShowcase from './components/border-widths-showcase/BorderWidthsShowcase';
 export default MdxLayout;
