+++ theme/doc/components/colors-showcase/ColorsShowcase.tsx
@@ -1,6 +1,5 @@
 import { groupColors } from '../../utils/group-colors';
-import React from 'react';
-import { ReactElement } from 'react';
+import React, { ReactElement } from 'react';
 import ColorSample from '../color-sample/ColorSample';
 import { GroupLabel, Samples } from './ColorsShowcase.styles';
 import { useTheme } from '../../../../theme-provider';
