+++ theme/doc/fonts.mdx
@@ -1,7 +1,7 @@
 import { mdx } from '@mdx-js/react';
 import { MdxLayout } from '~/layout';
 import FontsShowcase from './components/fonts-showcase/FontsShowcase';
-import { lightBaseTheme } from '../src/theme';
+import { lightBaseTheme } from '../src';
 export default MdxLayout;
 
 # FontsShowcase
