+++ theme/doc/space.mdx
@@ -1,7 +1,7 @@
 import { mdx } from '@mdx-js/react';
 import { SystemUIShowcases } from '@divriots/dockit-react/systemui-showcases';
 import { MdxLayout } from '~/layout';
-  import { lightBaseTheme } from '../src/theme';
+import { lightBaseTheme } from '../src';
 export default MdxLayout;
 
 # Space
